+++ src/services/tag/connections.ts
@@ -65,10 +65,13 @@ import { RawAwsElasticSearchDomain } from '../elasticSearchDomain/data'
 import { RawAwsSystemsManagerDocument } from '../systemsManagerDocument/data'
 import { RawAwsRdsClusterSnapshot } from '../rdsClusterSnapshot/data'
 import { RawAwsInstanceProfile } from '../iamInstanceProfile/data'
+import { RawAwsVpcEndpoint } from '../vpcEndpoint/data'
 import { RawAwsApiGatewayHttpApi } from '../apiGatewayHttpApi/data'
 import { RawAwsApiGatewayDomainName } from '../apiGatewayDomainName/data'
 import { RawAwsAnalyzerSummary } from '../iamAccessAnalyzer/data'
 import { RawAwsManagedPrefixList } from '../managedPrefixList/data'
+import { RawAwsTransitGatewayRouteTable } from '../transitGatewayRouteTable/data'
+import { RawAwsVpcPeeringConnection } from '../vpcPeeringConnection/data'
 
 const findServiceInstancesWithTag = (tag: any, service: any): any => {
   const { id } = tag
@@ -1777,6 +1780,84 @@ export default ({
             resourceType: services.apiGatewayDomainName,
             relation: 'child',
             field: 'apiGatewayDomainName',
+          })
+        }
+      }
+    }
+
+    /**
+     * Find related Vpc Endpoints
+     */
+    const vpcEndpoints: {
+      name: string
+      data: { [property: string]: RawAwsVpcEndpoint[] }
+    } = data.find(({ name }) => name === services.vpcEndpoint)
+    if (vpcEndpoints?.data?.[region]) {
+      const dataAtRegion = findServiceInstancesWithTag(
+        tag,
+        vpcEndpoints.data[region]
+      )
+      if (!isEmpty(dataAtRegion)) {
+        for (const instance of dataAtRegion) {
+          const { VpcEndpointId: id }: RawAwsVpcEndpoint = instance
+
+          connections.push({
+            id,
+            resourceType: services.vpcEndpoint,
+            relation: 'child',
+            field: 'vpcEndpoints',
+          })
+        }
+      }
+    }
+
+    /**
+     * Find related Vpc Peering Connections
+     */
+     const vpcPeeringConnections: {
+      name: string
+      data: { [property: string]: RawAwsVpcPeeringConnection[] }
+    } = data.find(({ name }) => name === services.vpcPeeringConnection)
+    if (vpcPeeringConnections?.data?.[region]) {
+      const dataAtRegion = findServiceInstancesWithTag(
+        tag,
+        vpcPeeringConnections.data[region]
+      )
+      if (!isEmpty(dataAtRegion)) {
+        for (const instance of dataAtRegion) {
+          const { VpcPeeringConnectionId: id }: RawAwsVpcPeeringConnection = instance
+
+          connections.push({
+            id,
+            resourceType: services.vpcPeeringConnection,
+            relation: 'child',
+            field: 'vpcPeeringConnections',
+          })
+        }
+      }
+    }
+
+    /**
+     * Find related Transit Gateway Route tables
+     */
+     const transitGatewayRouteTables: {
+      name: string
+      data: { [property: string]: RawAwsTransitGatewayRouteTable[] }
+    } = data.find(({ name }) => name === services.transitGatewayRouteTable)
+    if (transitGatewayRouteTables?.data?.[region]) {
+      const dataAtRegion = findServiceInstancesWithTag(
+        tag,
+        transitGatewayRouteTables.data[region]
+      )
+      if (!isEmpty(dataAtRegion)) {
+        for (const instance of dataAtRegion) {
+          const { TransitGatewayRouteTableId: id }: RawAwsTransitGatewayRouteTable = instance
+
+          connections.push({
+            id,
+            resourceType: services.transitGatewayRouteTable,
+            relation: 'child',
+            field: 'transitGatewayRouteTables',
           })
         }
       }