+++ src/services/vpcEndpoint/index.ts
@@ -2,12 +2,15 @@ import { Service } from '@cloudgraph/sdk'
 import BaseService from '../base'
 import format from './format'
 import getData from './data'
+import getConnections from './connections'
 import mutation from './mutation'
 
-export default class TransitGateway extends BaseService implements Service {
+export default class VpcEndpoint extends BaseService implements Service {
   format = format.bind(this)
 
   getData = getData.bind(this)
+
+  getConnections = getConnections.bind(this)
 
   mutation = mutation
 }